[hotfix][tests] Move test checkpoint component classes to upper level
